Add index file to deposit module with exports, refractor code
