+++ src/renderer/services/deposit/index.ts
@@ -1,0 +1,7 @@
+/**
+ * @module deposit
+ */
+
+export {DepositTx} from "./DepositTx";
+export {generateDeposit, functionSignatureFromABI} from "./utils";
+export {ITx} from "./types";
