+++ src/renderer/services/deposit/utils.ts
@@ -1,3 +1,12 @@
+import {Keypair as KeyPair} from "@chainsafe/bls/lib/keypair";
+import {BLSPubkey as BLSPubKey, DepositData} from "@chainsafe/eth2.0-types";
+import {createHash} from "crypto";
+import {signingRoot} from "@chainsafe/ssz";
+import {config} from "@chainsafe/eth2.0-config/lib/presets/mainnet";
+import {DEPOSIT_AMOUNT, DEPOSIT_DOMAIN} from "./constants";
+import {G2point} from "@chainsafe/bls/lib/helpers/g2point";
+import {EthConverter} from "../utils/crypto-utils";
+
 /**
  * Generate function signature from ABI object.
  *
@@ -22,4 +31,37 @@ export function functionSignatureFromABI(rawAbi: (string | any)[] | string, func
         }
     }
     return hasFunction ? `${functionName}(${inputs.join(",")})` : "";
+}
+
+/**
+ * Generate deposit params.
+ *
+ * @param signingKey
+ * @param withdrawalPubKey
+ *
+ * @return instance of ${DepositData}
+ */
+export function generateDeposit(signingKey: KeyPair, withdrawalPubKey: BLSPubKey): DepositData {
+    // signing public key
+    const publicKey = signingKey.publicKey.toBytesCompressed();
+    // BLS_WITHDRAWAL_PREFIX + hash(withdrawal_pubkey)[1:]
+    const withdrawalCredentials = Buffer.concat([
+        Buffer.alloc(1),
+        createHash("sha256").update(withdrawalPubKey).digest().subarray(1)
+    ]);
+    // define DepositData
+    const depositData: DepositData = {
+        pubkey: publicKey,
+        withdrawalCredentials: withdrawalCredentials,
+        amount: EthConverter.toGwei(DEPOSIT_AMOUNT),
+        signature: Buffer.alloc(0)
+    };
+    // calculate root
+    const root = signingRoot(depositData, config.types.DepositData);
+    // sign calculated root
+    depositData.signature = signingKey.privateKey.signMessage(
+        root,
+        DEPOSIT_DOMAIN
+    ).toBytesCompressed();
+    return depositData;
 }
